+++ javascripts/init_game.js
@@ -1,42 +1,3 @@
-/*Offline Progress*/
-function simulateTime(seconds, actual, testing) {
-	if (seconds > 10) {
-		document.getElementById("offlinePopup").style.display = "block";
-		document.getElementById("offlineProgress").innerText = "Simulating " + seconds + " seconds of progress.";
-	}
-	let ticks = seconds * 20;
-	let tickInterval = 50;
-	if (ticks > 1000 & !actual) {
-		tickInterval += (ticks - 1000) / 20;
-		ticks = 1000;
-	}
-	let start = Object.assign({}, player);
-	for (let complete = 0; complete < ticks; complete++) {
-		updateGame(tickInterval);
-		player.lastUpdate = Date.now();
-	}
-	player.time += seconds * 1000;
-	player.meltdown.time += seconds * 1000;
-	let offlinePopup = "While you were away, "
-	if (player.energy.gt(start.energy)) {
-		offlinePopup += "your energy increased by " + notation(player.energy.log10() - start.energy.log10()) + " Orders of Magnitude.";
-	}
-	if (offlinePopup == "While you were away, ") {
-		offlinePopup += "nothing happened...";
-	}
-	if (seconds > 1) {
-		document.getElementById("offlinePopup").style.display = "none";
-	}
-	if (seconds > 1000) {
-		document.getElementById("offlinePopup").style.display = "block";
-		document.getElementById("offlineProgress").innerText = offlinePopup;
-	}
-}
-
-function closeOfflineProgress() {
-	document.getElementById("offlinePopup").style.display = "none";
-}
-
 function init_game() {
 	loadSave();
 	document.getElementById("production").style.display = "none";
@@ -44,6 +5,7 @@ function init_game() {
 	document.getElementById("mines").style.display = "none";
 	document.getElementById("reactors").style.display = "none";
 	document.getElementById("statistics").style.display = "none";
+	document.getElementById("achievements").style.display = "none";
 	document.getElementById("options").style.display = "none";
 	document.getElementById("nanite").style.display = "none";
 	showNaviTab(player.navigation.naviTab);
@@ -52,6 +14,7 @@ function init_game() {
 	if (Date.now() > player.lastUpdate + 1000) {
 		simulateTime((Date.now() - player.lastUpdate) / 1000);
 	}
+	scrollNextNENR();
 }